Switched to typed structs from base
Added compiler flags to use the new typed structs from base
instead of the deprecated ones

+++ devgpioApp/src/GpioIntHandler.cpp
@@ -84,7 +84,7 @@ void GpioIntHandler::run() {
 
     for( auto r : _recs ) {
       struct gpio_v2_line_event event;
-      int rtn = read( r->fd, &event, sizeof(event));
+      ssize_t rtn = read( r->fd, &event, sizeof(event));
       if( -1 == rtn ) {
         if( errno == -EAGAIN ) {
           // nothing to read
